Se actualizan las clases de servicio para cambiar el tipo de dato de los identificadores de Long a Integer, mejorando la consistencia en la gestión de identificadores.

+++ back_local/src/main/java/com/portfolioweb/aj/Servicio/SExperiencia.java
@@ -122,7 +122,7 @@ public class SExperiencia {
             return Optional.of("La organizacion no existe");
         }
         if (dto.getHabilidadesIds() != null) {
-            for (Long habilidadId : dto.getHabilidadesIds()) {
+            for (Integer habilidadId : dto.getHabilidadesIds()) {
                 if (habilidadId == null || !rHabilidad.existsById(habilidadId)) {
                     return Optional.of("Una o mas habilidades no existen");
                 }
@@ -150,7 +150,7 @@ public class SExperiencia {
         return experiencia;
     }
 
-    private List<Habilidad> resolveHabilidades(List<Long> habilidadesIds) {
+    private List<Habilidad> resolveHabilidades(List<Integer> habilidadesIds) {
         if (habilidadesIds == null || habilidadesIds.isEmpty()) {
             return new ArrayList<>();
         }
