+++ back_local/src/main/java/com/portfolioweb/aj/Servicio/Seducacion.java
@@ -103,7 +103,7 @@ public class SEducacion {
             return Optional.of("La organizacion no existe");
         }
         if (dto.getHabilidadesIds() != null) {
-            for (Long habilidadId : dto.getHabilidadesIds()) {
+            for (Integer habilidadId : dto.getHabilidadesIds()) {
                 if (habilidadId == null || !rHabilidad.existsById(habilidadId)) {
                     return Optional.of("Una o mas habilidades no existen");
                 }
@@ -135,7 +135,7 @@ public class SEducacion {
         return educacion;
     }
 
-    private List<Habilidad> resolveHabilidades(List<Long> habilidadesIds) {
+    private List<Habilidad> resolveHabilidades(List<Integer> habilidadesIds) {
         if (habilidadesIds == null || habilidadesIds.isEmpty()) {
             return new ArrayList<>();
         }
